+++ lib/kitten/configuration.rb
@@ -188,6 +188,7 @@ module Kitten
             { path: 'components/cards/simple-card', name: 'Simple card' },
             { path: 'components/cards/project-card', name: 'Project card' },
             { path: 'components/cards/article-card', name: 'Article card' },
+            { path: 'components/cards/backer-card', name: 'Backer card' },
             {
               path: 'components/cards/crowdfunding-card',
               name: 'Crowdfunding card',