+++ spec/dummy/client/entries/app-kitten.js
@@ -42,6 +42,7 @@ import { KarlButtonIconFilter } from 'kitten/karl/buttons/button-filter'
 
 // Cards
 import { ArticleCard } from 'kitten/components/cards/article-card'
+import { BackerCard } from 'kitten/components/cards/backer-card'
 import { CrowdfundingCard } from 'kitten/components/cards/crowdfunding-card'
 import { ProjectCard } from 'kitten/components/cards/project-card'
 import { ProjectCreatorCard }
@@ -355,6 +356,7 @@ ReactOnRails.register({
 
   // Cards
   ArticleCard,
+  BackerCard,
   CrowdfundingCard,
   ProjectCard,
   ProjectCreatorCard,